FEAT - Exibindo bonificação por grupos de empreendimentos
