+++ src/app/BonificacoesGeradas/bonificacoesgeradas.module.css
@@ -4,6 +4,7 @@
 
 .table {
   margin-top: 18px;
+  height: 90vh;
 }
 
 .tableHeadCell {
